feat: implementa like, repost e save

+++ app/views.py
@@ -3,16 +3,32 @@ from django.contrib.auth.decorators import login_required
 from django.urls import reverse_lazy
 from django.views import generic
 from app.forms import CustomUserCreationForm
-from app.models import Post
+from app.models import Post, Like, Repost, Save
 
 @login_required(login_url='login')
 def home(request):
+    posts = None
+
     if request.method == 'GET':
         posts = Post.objects.order_by('-created_at')
-        context = {'posts': posts}
-        # for post in posts:
-        #     print('\033[0;34m'+post.content+'\033[m')
+        for post in posts:
+            try:
+                like = Like.objects.get(author = request.user, post = post)
+                post.liked_by_authenticated_user = True
+            except Like.DoesNotExist:
+                post.liked_by_authenticated_user = False
+            try:
+                repost = Repost.objects.get(author = request.user, post = post)
+                post.reposted_by_authenticated_user = True
+            except Repost.DoesNotExist:
+                post.reposted_by_authenticated_user = False
+            try:
+                save = Save.objects.get(author = request.user, post = post)
+                post.saved_by_authenticated_user = True
+            except Save.DoesNotExist:
+                post.saved_by_authenticated_user = False
 
+    context = {'posts': posts}
     return render(request, 'home.html', context)
 
 class SignUpView(generic.CreateView):
@@ -25,7 +41,7 @@ def create_post(request):
     if request.method == 'POST':
         content = request.POST.get('content')
 
-        novo_post = Post.objects.create(
+        Post.objects.create(
             author = request.user,
             content = content,
         )
@@ -34,6 +50,52 @@ def create_post(request):
 @login_required(login_url='login')
 def delete_post(request, post_id):
     post = get_object_or_404(Post, id = post_id)
-    post.delete()
+    if request.user == post.author:
+        post.delete()
     # print('\033[0;31mO post "'+post.content+'" foi apagado\033[m')
+    return redirect('home')
+
+@login_required(login_url='login')
+def like(request, post_id):
+    post = get_object_or_404(Post, id = post_id)
+    try:
+        like = Like.objects.get(author = request.user, post = post)
+        like.delete()
+        post.likes -= 1
+    except Like.DoesNotExist:
+        Like.objects.create(
+            author = request.user,
+            post = post,
+        )
+        post.likes += 1
+    
+    post.save()
+    return redirect('home')
+
+@login_required(login_url='login')
+def repost(request, post_id):
+    post = get_object_or_404(Post, id = post_id)
+    try:
+        repost = Repost.objects.get(author = request.user, post = post)
+        repost.delete()
+    except Repost.DoesNotExist:
+        Repost.objects.create(
+            author = request.user,
+            post = post,
+        )
+    
+    return redirect('home')
+
+@login_required(login_url='login')
+def save(request, post_id):
+    post = get_object_or_404(Post, id = post_id)
+    try:
+        save = Save.objects.get(author = request.user, post = post)
+        save.delete()
+    except Save.DoesNotExist:
+        Save.objects.create(
+            author = request.user,
+            post = post,
+        )
+    
     return redirect('home')
